relocate Bodu.Globalization.Calendar.DependencyInjection contract and KAT out of Bodu.Test
One contract base and one KAT record were consumed only by
Bodu.Globalization.Calendar.DependencyInjection.Test. Both move
alongside their consumer per the single-consumer rule. The
ServiceRegistrationKat that the earlier exploration flagged as
potentially orphaned was confirmed to be in use (the contract base
itself consumes it), so it relocates instead of being deleted.

Moved to Bodu.Globalization.Calendar.DependencyInjection/test/
Globalization.Calendar.DependencyInjection/Contracts/ (namespace
Bodu.Globalization.Calendar.DependencyInjection.Contracts to match
NotableDateServiceLifetimeContractTests):

- ServiceLifetimeContractTests
- ServiceRegistrationKat

NotableDateServiceLifetimeContractTests: removed `using Bodu.Test.
Contracts;` and `using Bodu.Test.Kat;` (both types now in same
namespace as the subclass).

Verified: dotnet build passes 0 errors;
dotnet test --settings bvt.runsettings passes 91/91.
